All Armor and Mask Items have had their file name adjusted to properly include spacing, multiple advancements and recipes have been changed to reflect this. Sky Ruins advancement has been fixed to now target Chasms. Fierce Deity Sword, Master Sword, and Magic Swords now properly implement SwingActionItem in order to shoot sword beams. Master Sword items have been condensed and renamed to properly indicate which variant they are. New Enchantment Sounds for upgrading the Master Sword have been implemented.
MASSIVE ORGANIZATION OVERHAUL!!
ModBlocks, Entities, Items, Armor types, etc have all been organized with propery spacing and locations for all items.

+++ src/main/java/com/kamth/zeldamod/entity/projectile/magic/SwordBeam.java
@@ -18,7 +18,7 @@ import net.minecraftforge.network.NetworkHooks;
 import org.jetbrains.annotations.NotNull;
 
 public class SwordBeam extends AbstractArrow {
-    private static final double BASE_DAMAGE = 3.0D;
+    private static final float BASE_DAMAGE = 3;
     public SwordBeam(EntityType<? extends AbstractArrow> pEntityType, Level pLevel) {
         super(pEntityType, pLevel);
     }
